drf/requirements.txt
django==3.2.9
djangorestframework==3.12.4
markdown==3.3.4
django-filter==21.1
django-cors-headers==3.10.0
